+++ packages/analytics-js/src/components/pluginsManager/types.ts
@@ -14,5 +14,18 @@ export interface IPluginsManager {
   register(plugins: ExtensionPlugin[]): void;
 }
 
-export { PluginName } from '@rudderstack/analytics-js-plugins/types/common';
+export enum PluginName {
+  BeaconQueue = 'BeaconQueue',
+  ConsentManager = 'ConsentManager',
+  DeviceModeDestinations = 'DeviceModeDestinations',
+  DeviceModeTransformation = 'DeviceModeTransformation',
+  ErrorReporting = 'ErrorReporting',
+  ExternalAnonymousId = 'ExternalAnonymousId',
+  GoogleLinker = 'GoogleLinker',
+  NativeDestinationQueue = 'NativeDestinationQueue',
+  StorageEncryption = 'StorageEncryption',
+  StorageEncryptionLegacy = 'StorageEncryptionLegacy',
+  XhrQueue = 'XhrQueue',
+}
+
 export type PluginMap<T = ExtensionPlugin> = Record<string, () => T>;
